Not done, but gotta sleep

+++ src/main/java/mandatory/two/controller/CourseController.java
@@ -1,5 +1,6 @@
 package mandatory.two.controller;
 
+import mandatory.two.helper.SessionHelper;
 import mandatory.two.model.*;
 import mandatory.two.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,6 @@ import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,7 +35,7 @@ public class CourseController {
     private Long editId;
 
     @GetMapping("/course")
-    public String courses(Model model){
+    public String courses(Model model, HttpServletRequest request) {
         List<Course> courseList = courseRepo.findAll();
         //ArrayList<StudyProgramme> studyList = (ArrayList<StudyProgramme>) studyRepo.findAll();
         //System.out.println(studyList);
@@ -43,80 +43,97 @@ public class CourseController {
         //Course c = courseRepository.findByEcts(2);
         //System.out.println(c);
         model.addAttribute("courses", courseList);
-        return "courses";
+        return SessionHelper.redirectTeacher(request, "courses");
     }
 
     @GetMapping("/course/delete/{id}")
-    public String deleteCourse(@PathVariable Long id){
-        courseRepo.deleteById(id);
+    public String deleteCourse(@PathVariable Long id, HttpServletRequest request) {
+        if (SessionHelper.isTeacher(request)) {
+            courseRepo.deleteById(id);
+        }
         return "redirect:/course";
     }
+
     @GetMapping("/course/edit/{id}")
-    public String editCourseView(Model model, @PathVariable Long id){
+    public String editCourseView(Model model, @PathVariable Long id, HttpServletRequest request) {
         Optional<Course> courseOptional = courseRepo.findById(id);
         Course course = courseOptional.get();
         editId = course.getId();
         model.addAttribute("course", course);
         model.addAttribute("teacher", teacherRepo.findAll());
         model.addAttribute("studyprogramme", studyRepo.findAll());
-        return "editCourse";
+        return SessionHelper.redirectTeacher(request, "editCourse");
     }
-    @PostMapping("/course/edit")
-    public String editCourse(@ModelAttribute Course course) {
-        course.setId(editId);
-        editId = null;
-        courseRepo.save(course);
 
+    @PostMapping("/course/edit")
+    public String editCourse(@ModelAttribute Course course, HttpServletRequest request) {
+        if (SessionHelper.isTeacher(request)) {
+            course.setId(editId);
+            editId = null;
+            courseRepo.save(course);
+        }
         return "redirect:/course";
     }
 
     @GetMapping("/course/create")
-    public String createCourse(Model model){
+    public String createCourse(Model model, HttpServletRequest request) {
         model.addAttribute("teacher", teacherRepo.findAll());
         model.addAttribute("studyprogramme", studyRepo.findAll());
         model.addAttribute("course", new Course());
-        return "createCourse";
+        return SessionHelper.redirectTeacher(request, "createCourse");
     }
+
     @PostMapping("course/create")
-    public String createCourse(@ModelAttribute Course course){
+    public String createCourse(@ModelAttribute Course course, HttpServletRequest request) {
         //System.out.println(course.toString());
-        courseRepo.save(course);
+        if (SessionHelper.isTeacher(request)) {
+            courseRepo.save(course);
+        }
         return "redirect:/course";
     }
 
     @GetMapping("/course/join/{id}")
     public ModelAndView joinCourse(@PathVariable Long id, HttpServletRequest request) {
+        if (SessionHelper.isStudent(request)) {
 
-        HttpSession session = request.getSession();
-        Student student = (Student) session.getAttribute("user");
+            HttpSession session = request.getSession();
+            Student student = (Student) session.getAttribute("user");
 
-        Optional<Course> optionalCourse = courseRepo.findById(id);
-        Course course = optionalCourse.get();
+            Optional<Course> optionalCourse = courseRepo.findById(id);
+            Course course = optionalCourse.get();
 
-        course.addStudentToWaitingList(student);
+            course.addStudentToWaitingList(student);
 
-        courseRepo.save(course);
+            courseRepo.save(course);
 
-        ModelAndView mav = (ModelAndView) session.getAttribute("lastView");
+            ModelAndView mav = (ModelAndView) session.getAttribute("lastView");
 
-        return mav;
+            return mav;
+        } else {
+            return new ModelAndView("all/forbidden");
+        }
     }
 
     @GetMapping("/course/showmore")
-    public String showMore(@RequestParam Long id, Model model){
-        Optional<Course> optionalCourse = courseRepo.findById(id);
-        Course course = optionalCourse.get();
-        model.addAttribute("course", course);
-        return "courseInfo";
+    public String showMore(@RequestParam Long id, Model model, HttpServletRequest request) {
+        if (SessionHelper.isSessionValid(request)) {
+            Optional<Course> optionalCourse = courseRepo.findById(id);
+            Course course = optionalCourse.get();
+            model.addAttribute("course", course);
+            model.addAttribute("userType", request.getSession().getAttribute("user").getClass().getSimpleName());
+            return "all/courseInfo";
+        } else {
+            return "all/forbidden";
+        }
     }
 
     @GetMapping("/course/students/{id}")
-    public String courseStudents(@PathVariable Long id, Model model){
+    public String courseStudents(@PathVariable Long id, Model model, HttpServletRequest request) {
         Optional<Course> optionalCourse = courseRepo.findById(id);
         Course course = optionalCourse.get();
         model.addAttribute("student", studentRepo.findAll());
         model.addAttribute("course", course);
-        return "courseStudents";
+        return SessionHelper.redirectTeacher(request, "courseStudents");
     }
 
 }
